Allow PresenceStore to be instansiated on workers
